+++ Wafer/Wafer.UI.Direct2D/Factories/ITextLayoutFactory.cs
@@ -2,7 +2,7 @@ using SharpDX.DirectWrite;
 
 namespace Wafer.UI.Direct2D.Factories {
     public interface ITextLayoutFactory {
-        TextLayout CreateTextLayout(TextFormat format, string text, int maximumWidth, int maximumHeight);
+        TextLayout CreateTextLayout(TextFormat format, string text, float maximumWidth, float maximumHeight);
 
         void DisposeItems();
     }
